move building use and building parser into one file, small fixes
