+++ app/src/commonTest/kotlin/de/westnordost/streetcomplete/osm/building/BuildingTypeParserKtTest.kt
@@ -3,6 +3,7 @@ package de.westnordost.streetcomplete.osm.building
 import de.westnordost.streetcomplete.osm.building.BuildingType.*
 import kotlin.test.Test
 import kotlin.test.assertEquals
+import kotlin.test.assertNull
 
 class BuildingTypeParserKtTest {
 
@@ -44,5 +45,42 @@ class BuildingTypeParserKtTest {
         assertEquals(HISTORIC, createBuildingType(mapOf("historic" to "yes")))
         assertEquals(HISTORIC, createBuildingType(mapOf("building" to "yes", "historic" to "yes")))
         assertEquals(APARTMENTS, createBuildingType(mapOf("building" to "apartments", "historic" to "yes")))
+    }
+
+    @Test fun `parse missing building use`() {
+        assertNull(createBuildingUseType(mapOf()))
+        assertNull(createBuildingUseType(mapOf("building" to "office")))
+    }
+
+    @Test fun `requires building key`() {
+        assertNull(createBuildingUseType(mapOf("building:use" to "tower")))
+    }
+
+    @Test fun `parse building use`() {
+        assertEquals(OFFICE, createBuildingUseType(mapOf("building" to "commercial", "building:use" to "office")))
+        assertEquals(APARTMENTS, createBuildingUseType(mapOf("building" to "residential", "building:use" to "apartments")))
+    }
+
+    @Test fun `parse building use aliases`() {
+        assertEquals(FARM_AUXILIARY, createBuildingUseType(mapOf("building:use" to "livestock")))
+        assertEquals(RELIGIOUS, createBuildingUseType(mapOf("building:use" to "convent")))
+    }
+
+    @Test fun `parse houses uses with alternative tagging scheme`() {
+        assertEquals(HOUSE, createBuildingUseType(mapOf("building:use" to "house")))
+        assertEquals(BUNGALOW, createBuildingUseType(mapOf("building:use" to "house", "house" to "bungalow")))
+        assertEquals(DETACHED, createBuildingUseType(mapOf("building:use" to "house", "house" to "detached")))
+    }
+
+    @Test fun `parse excluded building uses`() {
+        assertNull(createBuildingUseType(mapOf("building:use" to "yes")))
+    }
+
+    @Test fun `parse unsupported building use`() {
+        assertNull(createBuildingUseType(mapOf("building:use" to "something")))
+    }
+
+    @Test fun `parse invalid building use`() {
+        assertNull(createBuildingUseType(mapOf("building:use" to "glasshouse")))
     }
 }
